backend/requirements.txt
asgiref==3.9.1
boto3==1.40.4
botocore==1.40.4
certifi==2025.7.14
charset-normalizer==3.4.2
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==4.2.23
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
stripe==12.3.0
typing_extensions==4.14.1
urllib3==1.26.20
whitenoise==6.9.0
